Stworzona osobna klasa implementujaca taxPolicy
