+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/invoicing/BookKeeper.java
@@ -12,22 +12,19 @@
  */
 package pl.com.bottega.ecommerce.sales.domain.invoicing;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import pl.com.bottega.ecommerce.canonicalmodel.publishedlanguage.ClientData;
-import pl.com.bottega.ecommerce.canonicalmodel.publishedlanguage.Id;
-import pl.com.bottega.ecommerce.sales.domain.productscatalog.ProductType;
 import pl.com.bottega.ecommerce.sharedkernel.Money;
 
-public class BookKeeper implements TaxPolicy{
+public class BookKeeper{
 
-    public Invoice issuance(ClientData client, List<RequestItem> items, InvoiceFactory invoiceFactory) {
+    public Invoice issuance(ClientData client, List<RequestItem> items, InvoiceFactory invoiceFactory, TaxPolicy taxPolicy) {
         Invoice invoice=invoiceFactory.createInvoice(client);
         
         for (RequestItem item : items) {
             Money net = item.getTotalCost();
-            Tax tax =createTax(item.getProductData().getType(), net);
+            Tax tax=taxPolicy.createTax(item.getProductData().getType(), net);
 
             InvoiceLine invoiceLine = new InvoiceLine(item.getProductData(), item.getQuantity(), net, tax);
             invoice.addItem(invoiceLine);
@@ -35,29 +32,4 @@ public class BookKeeper implements TaxPolicy{
 
         return invoice;
     }
-
-    @Override
-    public Tax createTax(ProductType productType, Money net) {
-        BigDecimal ratio;
-        String desc;
-        switch (productType) {
-            case DRUG:
-                ratio = BigDecimal.valueOf(0.05);
-                desc = "5% (D)";
-                break;
-            case FOOD:
-                ratio = BigDecimal.valueOf(0.07);
-                desc = "7% (F)";
-                break;
-            case STANDARD:
-                ratio = BigDecimal.valueOf(0.23);
-                desc = "23%";
-                break;
-            default:
-                throw new IllegalArgumentException(productType + " not handled");
-        }
-        Money taxValue = net.multiplyBy(ratio);
-        return new Tax(taxValue, desc);
-    }
-
 }
